+++ PFM.Website.Reboot/src/PFM.Website/Configurations/ApplicationSettings.cs
@@ -1,4 +1,7 @@
 ﻿using System;
+using PFM.Website.Monitoring.Metrics;
+using PFM.Website.Monitoring.Tracing;
+
 namespace PFM.Website.Configurations
 {
 	public class ApplicationSettings
@@ -8,9 +11,7 @@ namespace PFM.Website.Configurations
         public string AwsRegion { get; init; } = "eu-west-2";
         public string AwsEndpointUrl { get; init; } = string.Empty;
         public PfmApiOptions PfmApiOptions { get; init; } = new ();
-        
         public MetricsOptions MetricsOptions { get; init; } = new ();
-        
         public TracingOptions TracingOptions { get; init; } = new ();
 	}
 
@@ -18,16 +19,6 @@ namespace PFM.Website.Configurations
 	{
 		public bool Enabled { get; init; } = false;
 		public string EndpointUrl { get; init; } = string.Empty;
-	}
-
-	public class MetricsOptions
-	{
-		public bool Debug { get; init; } = false;
-	}
-	
-	public class TracingOptions
-	{
-		public bool Debug { get; init; } = false;
 	}
 }
