+++ PFM.Website.Reboot/src/PFM.Website/Monitoring/Metrics/MetricsOptions.cs
@@ -1,0 +1,6 @@
+namespace PFM.Website.Monitoring.Metrics;
+
+public class MetricsOptions
+{
+    public bool Debug { get; init; } = false;
+}
