+++ PFM.Website.Reboot/src/PFM.Website/Monitoring/Tracing/TracingOptions.cs
@@ -1,0 +1,6 @@
+namespace PFM.Website.Monitoring.Tracing;
+
+public class TracingOptions
+{
+    public bool Debug { get; init; } = false;
+}
