+++ PFM.Website.Reboot/src/PFM.Website/Program.cs
@@ -1,6 +1,9 @@
 ﻿using Amazon;
 using Amazon.S3;
 using PFM.Website.Configurations;
+using PFM.Website.Monitoring.Logging;
+using PFM.Website.Monitoring.Metrics;
+using PFM.Website.Monitoring.Tracing;
 using PFM.Website.Persistence;
 using PFM.Website.Persistence.Implementations;
 using PFM.Website.Services;
@@ -62,7 +65,9 @@ builder.Services.AddTransient<AuthHeaderHandler>();
 builder.Services
     .AddAuth(builder.Configuration)
     .AddObjectMapper()
-    .AddMonitoring(builder.Configuration, builder.Environment, appSettings)
+    .ConfigureLogging(builder.Configuration, builder.Environment)
+    .ConfigureTracing(appSettings.TracingOptions)
+    .ConfigureMetrics(appSettings.MetricsOptions)
     .AddPfmApi(builder.Configuration, builder.Environment.EnvironmentName != "Production");
 
 var app = builder.Build();
