+++ PFM.Website.Reboot/src/PFM.Website/appsettings.json
@@ -40,7 +40,7 @@
       "Debug": false
     },
     "TracingOptions": {
-      "Debug": true
+      "Debug": false
     }
   }
 }
